+++ src/api/reddit/models/download_meta.rs
@@ -9,8 +9,8 @@ use super::download_status::DownloadStatus;
 pub struct DownloadMeta {
 	pub url: String,
 	pub subreddit_name: String,
-	pub image_height: usize,
-	pub image_width: usize,
+	pub image_height: u32,
+	pub image_width: u32,
 	pub post_link: String,
 	pub nsfw: bool,
 	pub filename: String,
